# Conflicts:
#	.idea/jpa-buddy.xml
#	.idea/misc.xml
#	pom.xml

+++ api/src/main/java/entities/devices/drives/Drive.java
@@ -11,7 +11,7 @@ public class Drive extends Hardware {
     @JsonProperty("deviceName")
     private String deviceName;
     @JsonProperty("temperature")
-    private long temperature;
+    private double temperature;
     @JsonProperty("partitions")
     private List<Partition> partitions;
 }
